dashboard/bim-istekleri/backend/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
firebase-admin==6.*
requests==2.*
openpyxl==3.*
